apps/ingest/pyproject.toml
[project]
name = "rag-ingest"
version = "0.1.0"
description = "Document ingestion worker for the Private RAG Accelerator"
requires-python = ">=3.12"
dependencies = [
    # Azure SDKs (managed identity, no shared keys)
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "azure-storage-queue>=12.12.0",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-search-documents>=11.6.0",
    "azure-cosmos>=4.9.0",
    
    # Azure OpenAI via openai SDK (for embeddings if not using integrated vectorization)
    "openai>=1.55.0",
    
    # Document processing
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "unstructured>=0.16.0",
    "pillow>=11.0.0",
    
    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    
    # Observability
    "azure-monitor-opentelemetry>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    # Shared test fixtures (tests/_shared/fixtures.py) sign synthetic Entra
    # JWTs to exercise the auth path future ingest endpoints will adopt.
    "python-jose[cryptography]>=3.3.0",
    "types-python-jose>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require deployed Azure resources",
    "integration_invnet: requires in-VNet runner with real Azure deps (Cosmos, Search, OpenAI, Storage). Skipped unless RUN_INVNET_TESTS=1.",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
